Move procedures in it's own module
This way, we can use it from bot.
