+++ jenkins_ghp/main.py
@@ -20,139 +20,15 @@ import inspect
 import logging
 import sys
 
+from jenkins_ghp import procedures
+
 from .bot import Bot
 from .cache import CACHE
-from .github import ApiError, ApiNotFoundError, GITHUB, cached_request
 from .jenkins import JENKINS
-from .repository import Branch, PullRequest, Repository
 from .settings import SETTINGS
-from .utils import match, retry
 
 
 logger = logging.getLogger('jenkins_ghp')
-
-
-class Procedures(object):
-    pr_filter = [p for p in str(SETTINGS.GHP_PR).split(',') if p]
-
-    @staticmethod
-    @retry(wait_fixed=15000)
-    def fetch_settings(repository):
-        try:
-            ghp_yml = GITHUB.fetch_file_contents(repository, '.github/ghp.yml')
-            logger.debug("Loading settings from .github/ghp.yml")
-        except ApiNotFoundError:
-            ghp_yml = None
-
-        collaborators = cached_request(GITHUB.repos(repository).collaborators)
-        branches = cached_request(
-            GITHUB.repos(repository).branches, protected='true',
-        )
-
-        repository.load_settings(
-            branches=branches,
-            collaborators=collaborators,
-            ghp_yml=ghp_yml,
-        )
-
-    @staticmethod
-    @retry(wait_fixed=15000)
-    def list_branches(repository):
-        branches = repository.SETTINGS.GHP_BRANCHES
-        if not branches:
-            logger.debug("No explicit branches configured for %s", repository)
-            return []
-
-        for branch in branches:
-            logger.debug("Search remote branch %s", branch)
-            try:
-                ref = cached_request(GITHUB.repos(repository).git(branch))
-            except ApiNotFoundError:
-                logger.warn("Branch %s not found in %s", branch, repository)
-                continue
-
-            sha = ref['object']['sha']
-            logger.debug("Fetching commit %s", sha[:7])
-            data = cached_request(GITHUB.repos(repository).commits(sha))
-            commit = data['commit']
-            branch = Branch(repository, ref, commit)
-            if branch.is_outdated:
-                logger.debug(
-                    'Skipping branch %s because older than %s weeks',
-                    branch, repository.SETTINGS.GHP_COMMIT_MAX_WEEKS,
-                )
-                continue
-            yield branch
-
-    @classmethod
-    @retry(wait_fixed=15000)
-    def list_pulls(cls, repository):
-        logger.debug("Querying GitHub for %s PR.", repository)
-        try:
-            pulls = cached_request(GITHUB.repos(repository).pulls)
-        except Exception:
-            logger.exception("Failed to list PR for %s.", repository)
-            return []
-
-        pulls_o = []
-        for data in pulls:
-            if not match(data['html_url'], cls.pr_filter):
-                logger.debug(
-                    "Skipping %s (%s).", data['html_url'], data['head']['ref'],
-                )
-            else:
-                pulls_o.append(PullRequest(repository, data))
-
-        for pr in reversed(sorted(pulls_o, key=PullRequest.sort_key)):
-            if pr.is_outdated:
-                logger.debug(
-                    'Skipping PR %s because older than %s weeks.',
-                    pr, SETTINGS.GHP_COMMIT_MAX_WEEKS,
-                )
-            else:
-                yield pr
-
-    @classmethod
-    def list_repositories(cls, fetch_settings=False):
-        repositories = {}
-        jobs = JENKINS.get_jobs()
-
-        env_repos = filter(None, SETTINGS.GHP_REPOSITORIES.split(' '))
-        for entry in env_repos:
-            repository, branches = (entry + ':').split(':', 1)
-            owner, name = repository.split('/')
-            repositories[repository] = Repository(owner, name)
-            logger.debug("Managing %s.", repository)
-
-        for job in jobs:
-            for remote in job.get_scm_url():
-                repository = Repository.from_remote(remote)
-                if repository not in repositories:
-                    logger.debug("Managing %s.", repository)
-                    repositories[repository] = repository
-                else:
-                    repository = repositories[repository]
-
-                logger.info("Managing %s.", job)
-                repository.jobs.append(job)
-                break
-            else:
-                logger.debug("Skipping %s, no GitHub repository.", job)
-
-        for repo in sorted(repositories.values(), key=str):
-            try:
-                if fetch_settings:
-                    Procedures.fetch_settings(repo)
-                yield repo
-            except ApiError as e:
-                logger.error("Failed to load %s settings: %r", repo, e)
-
-    @staticmethod
-    @retry(wait_fixed=15000)
-    def whoami():
-        user = cached_request(GITHUB.user)
-        logger.info("I'm @%s on GitHub.", user['login'])
-        return user['login']
 
 
 def loop(wrapped):
@@ -205,13 +81,13 @@ def check_queue(bot):
 @asyncio.coroutine
 def bot():
     """Poll GitHub to find something to do"""
-    Procedures.whoami()
+    procedures.whoami()
     bot = Bot(queue_empty=None)
 
-    for repository in Procedures.list_repositories(fetch_settings=True):
+    for repository in procedures.list_repositories(with_settings=True):
         logger.info("Working on %s.", repository)
 
-        for branch in Procedures.list_branches(repository):
+        for branch in procedures.list_branches(repository):
             try:
                 yield from check_queue(bot)
             except RestartLoop:
@@ -219,7 +95,7 @@ def bot():
                     break
             bot.run(branch)
 
-        for pr in Procedures.list_pulls(repository):
+        for pr in procedures.list_pulls(repository):
             try:
                 yield from check_queue(bot)
             except RestartLoop:
@@ -238,32 +114,32 @@ def bot():
 
 def list_jobs():
     """List managed jobs"""
-    for repository in Procedures.list_repositories():
+    for repository in procedures.list_repositories():
         for job in repository.jobs:
             print(job)
 
 
 def list_branches():
     """List branches to build"""
-    Procedures.whoami()
-    for repository in Procedures.list_repositories(fetch_settings=True):
+    procedures.whoami()
+    for repository in procedures.list_repositories(with_settings=True):
         logger.info("Working on %s.", repository)
-        for branch in Procedures.list_branches(repository):
+        for branch in procedures.list_branches(repository):
             print(branch)
 
 
 def list_pr():
     """List GitHub PR polled"""
-    Procedures.whoami()
-    for repository in Procedures.list_repositories(fetch_settings=True):
+    procedures.whoami()
+    for repository in procedures.list_repositories(with_settings=True):
         logger.info("Working on %s.", repository)
-        for pr in Procedures.list_pulls(repository):
+        for pr in procedures.list_pulls(repository):
             print(pr)
 
 
 def list_repositories():
     """List GitHub repositories tested by this Jenkins"""
-    for repository in Procedures.list_repositories():
+    for repository in procedures.list_repositories():
         print(repository)
